patch

+++ src/main/java/com/example/demo/Repository/ParkingInfoRepository.java
@@ -8,6 +8,6 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Transactional
-public interface ParkingInfoRepository extends JpaRepository<ParkingInfo, Long> {
+public interface ParkingInfoRepository extends JpaRepository<ParkingInfo, Integer> {
 
 }
